policies

+++ app/Http/Controllers/PortfolioController.php
@@ -37,6 +37,7 @@ class PortfolioController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Portfolio::class);
         $request->validate([
             "url" => "required",
             "name" => "required",
@@ -86,6 +87,7 @@ class PortfolioController extends Controller
      */
     public function update(Request $request, Portfolio $portfolio)
     {
+        $this->authorize('update', $portfolio);
         $request->validate([
             "url" => "required",
             "name" => "required",
@@ -112,6 +114,7 @@ class PortfolioController extends Controller
      */
     public function destroy(Portfolio $portfolio)
     {
+        $this->authorize('destroy', $portfolio);
         Storage::disk('public')->delete('img/'.$portfolio->url);
         $portfolio->delete();
         return redirect()->route('portfolios.index')->with('message', 'Elément supprimé avec succès');
